src/lib/decorators/firebase-user-delete.ts
import { getClassMethod, getClassName, addFirebaseFunction } from '../internal-methods';
import { FirebaseFunction, FirebaseTriggerType } from '../types';

/**
 * Decorator that adds the method to the Cloud Functions list that is
 * triggered when removing a user from Firebase Authentication
 */
export function onFirebaseUserDelete() {
  return (target: any, key: string) => {
    const firebaseFunction: FirebaseFunction = {
      className: getClassName(target),
      methodName: key,
      method: getClassMethod(target, key),
      trigger: FirebaseTriggerType.USER_DELETE,
    };
    addFirebaseFunction(firebaseFunction);
  };
}
